[Refactor] move all of Types about AboutMeInfo in AboutMeInfoType.ts file #89
