+++ app/src/lib/common/types/ComponentTypes/Contact/AboutMeInfoType.ts
@@ -1,0 +1,4 @@
+export interface AboutMeInfoPropsType {
+  title: string;
+  description: string;
+}
