+++ app/src/lib/components/Contact/ContactForm.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import Channels from '../Channels/Channels';
-import { ContactOptionPropsType } from '../../common/types/ComponentTypes/ContactType';
+import { ContactOptionPropsType } from '../../common/types/ComponentTypes/Contact/ContactType';
 
 const ContactForm = ({ contactOption }: ContactOptionPropsType) => {
   const { title, subTitle, buttonText, channels, email } = contactOption;
